refactor(mapper): move qr dialog to separate component
